move manuscript_type to common
