+++ ebl/common/domain/manuscript_type.py
