+++ ebl/corpus/domain/chapter_transformer.py
@@ -10,7 +10,7 @@ from ebl.corpus.domain.manuscript import (
     Manuscript,
     Siglum,
 )
-from ebl.corpus.domain.manuscript_type import ManuscriptType
+from ebl.common.domain.manuscript_type import ManuscriptType
 from ebl.common.domain.provenance import Provenance
 from ebl.transliteration.domain.line_transformer import LineTransformer
