+++ ebl/tests/factories/corpus.py
@@ -11,7 +11,7 @@ from ebl.corpus.domain.manuscript import (
     Manuscript,
     OldSiglum,
 )
-from ebl.corpus.domain.manuscript_type import ManuscriptType
+from ebl.common.domain.manuscript_type import ManuscriptType
 from ebl.common.domain.provenance import Provenance
 from ebl.corpus.domain.record import Author, AuthorRole, Record, Translator
 from ebl.corpus.domain.text import ChapterListing, Text
